Ruff: fix some SIM

+++ dojo/product/views.py
@@ -1,6 +1,7 @@
 # #  product
 import base64
 import calendar as tcalendar
+import contextlib
 import logging
 from collections import OrderedDict
 from datetime import date, datetime, timedelta
@@ -351,14 +352,11 @@ def identify_view(request):
         msg = 'invalid view, view must be "Endpoint" or "Finding"'
         raise ValueError(msg)
     else:
-        if get_data.get('finding__severity', None):
-            return 'Endpoint'
-        elif get_data.get('false_positive', None):
+        if get_data.get('finding__severity', None) or get_data.get('false_positive', None):
             return 'Endpoint'
     referer = request.META.get('HTTP_REFERER', None)
-    if referer:
-        if referer.find('type=Endpoint') > -1:
-            return 'Endpoint'
+    if referer and referer.find('type=Endpoint') > -1:
+        return 'Endpoint'
     return 'Finding'
 
 
@@ -610,13 +608,11 @@ def view_product_metrics(request, pid):
                 open_close_weekly[unix_timestamp] = {'closed': 0, 'open': 1, 'accepted': 0}
                 open_close_weekly[unix_timestamp]['week'] = html_date
 
-            if view == 'Finding':
-                severity = finding.get('severity')
-            elif view == 'Endpoint':
+            if view == 'Finding' or view == 'Endpoint':
                 severity = finding.get('severity')
 
             finding_age = calculate_finding_age(finding)
-            if open_objs_by_age.get(finding_age, None):
+            if open_objs_by_age.get(finding_age):
                 open_objs_by_age[finding_age] += 1
             else:
                 open_objs_by_age[finding_age] = 1
@@ -870,37 +866,36 @@ def new_product(request, ptid=None):
             success, jira_project_form = jira_helper.process_jira_project_form(request, product=product)
             error = not success
 
-            if get_system_setting('enable_github'):
-                if gform.is_valid():
-                    github_pkey = gform.save(commit=False)
-                    if github_pkey.git_conf is not None and github_pkey.git_project:
-                        github_pkey.product = product
-                        github_pkey.save()
-                        messages.add_message(request,
-                                             messages.SUCCESS,
-                                             _('GitHub information added successfully.'),
-                                             extra_tags='alert-success')
-                        # Create appropriate labels in the repo
-                        logger.info('Create label in repo: ' + github_pkey.git_project)
+            if get_system_setting('enable_github') and gform.is_valid():
+                github_pkey = gform.save(commit=False)
+                if github_pkey.git_conf is not None and github_pkey.git_project:
+                    github_pkey.product = product
+                    github_pkey.save()
+                    messages.add_message(request,
+                                         messages.SUCCESS,
+                                         _('GitHub information added successfully.'),
+                                         extra_tags='alert-success')
+                    # Create appropriate labels in the repo
+                    logger.info('Create label in repo: ' + github_pkey.git_project)
 
-                        description = _("This label is automatically applied to all issues created by DefectDojo")
-                        try:
-                            g = Github(github_pkey.git_conf.api_key)
-                            repo = g.get_repo(github_pkey.git_project)
-                            repo.create_label(name="security", color="FF0000",
-                                              description=description)
-                            repo.create_label(name="security / info", color="00FEFC",
-                                              description=description)
-                            repo.create_label(name="security / low", color="B7FE00",
-                                              description=description)
-                            repo.create_label(name="security / medium", color="FEFE00",
-                                              description=description)
-                            repo.create_label(name="security / high", color="FE9A00",
-                                              description=description)
-                            repo.create_label(name="security / critical", color="FE2200",
-                                              description=description)
-                        except:
-                            logger.info('Labels cannot be created - they may already exists')
+                    description = _("This label is automatically applied to all issues created by DefectDojo")
+                    try:
+                        g = Github(github_pkey.git_conf.api_key)
+                        repo = g.get_repo(github_pkey.git_project)
+                        repo.create_label(name="security", color="FF0000",
+                                          description=description)
+                        repo.create_label(name="security / info", color="00FEFC",
+                                          description=description)
+                        repo.create_label(name="security / low", color="B7FE00",
+                                          description=description)
+                        repo.create_label(name="security / medium", color="FEFE00",
+                                          description=description)
+                        repo.create_label(name="security / high", color="FE9A00",
+                                          description=description)
+                        repo.create_label(name="security / critical", color="FE2200",
+                                          description=description)
+                    except:
+                        logger.info('Labels cannot be created - they may already exists')
 
             if not error:
                 return HttpResponseRedirect(reverse('view_product', args=(product.id,)))
@@ -911,10 +906,7 @@ def new_product(request, ptid=None):
         if get_system_setting('enable_jira'):
             jira_project_form = JIRAProjectForm()
 
-        if get_system_setting('enable_github'):
-            gform = GITHUB_Product_Form()
-        else:
-            gform = None
+        gform = GITHUB_Product_Form() if get_system_setting('enable_github') else None
 
     add_breadcrumb(title=_("New Product"), top_level=False, request=request)
     return render(request, 'dojo/new_product.html',
@@ -961,10 +953,8 @@ def edit_product(request, pid):
             if get_system_setting('enable_github') and github_inst:
                 gform = GITHUB_Product_Form(request.POST, instance=github_inst)
                 # need to handle delete
-                try:
+                with contextlib.suppress(Exception):
                     gform.save()
-                except:
-                    pass
             elif get_system_setting('enable_github'):
                 gform = GITHUB_Product_Form(request.POST)
                 if gform.is_valid():
@@ -988,10 +978,7 @@ def edit_product(request, pid):
             jform = None
 
         if github_enabled:
-            if github_inst is not None:
-                gform = GITHUB_Product_Form(instance=github_inst)
-            else:
-                gform = GITHUB_Product_Form()
+            gform = GITHUB_Product_Form(instance=github_inst) if github_inst is not None else GITHUB_Product_Form()
         else:
             gform = None
 
@@ -1125,10 +1112,7 @@ def new_eng_for_app(request, pid, cicd=False):
             logger.debug('showing jira-epic-form')
             jira_epic_form = JIRAEngagementForm()
 
-    if cicd:
-        title = _('New CI/CD Engagement')
-    else:
-        title = _('New Interactive Engagement')
+    title = _('New CI/CD Engagement') if cicd else _('New Interactive Engagement')
 
     product_tab = Product_Tab(product, title=title, tab="engagements")
     return render(request, 'dojo/new_eng.html', {
@@ -1613,16 +1597,15 @@ def delete_engagement_presets(request, pid, eid):
     preset = get_object_or_404(Engagement_Presets, id=eid)
     form = DeleteEngagementPresetsForm(instance=preset)
 
-    if request.method == 'POST':
-        if 'id' in request.POST:
-            form = DeleteEngagementPresetsForm(request.POST, instance=preset)
-            if form.is_valid():
-                preset.delete()
-                messages.add_message(request,
-                                     messages.SUCCESS,
-                                     _('Engagement presets and engagement relationships removed.'),
-                                     extra_tags='alert-success')
-                return HttpResponseRedirect(reverse('engagement_presets', args=(pid,)))
+    if request.method == 'POST' and 'id' in request.POST:
+        form = DeleteEngagementPresetsForm(request.POST, instance=preset)
+        if form.is_valid():
+            preset.delete()
+            messages.add_message(request,
+                                 messages.SUCCESS,
+                                 _('Engagement presets and engagement relationships removed.'),
+                                 extra_tags='alert-success')
+            return HttpResponseRedirect(reverse('engagement_presets', args=(pid,)))
 
     collector = NestedObjects(using=DEFAULT_DB_ALIAS)
     collector.collect([preset])
